clean up gpu coordinate updater

+++ src/common/hist_util.cu
@@ -105,8 +105,8 @@ struct SketchContainer {
     // Initialize Sketches for this dmatrix
     sketches_.resize(info.num_col_);
 #pragma omp parallel for default(none) shared(info, param) schedule(static) \
-if (info.num_col_ > kOmpNumColsParallelizeLimit)
-    for (uint64_t icol = 0; icol < info.num_col_; ++icol) {
+if (info.num_col_ > kOmpNumColsParallelizeLimit)  // NOLINT
+    for (int icol = 0; icol < info.num_col_; ++icol) {  // NOLINT
       sketches_[icol].Init(info.num_row_, 1.0 / (8 * param.max_bin));
     }
   }
@@ -156,7 +156,7 @@ struct GPUSketcher {
         sketch_container_(sketch_container) {
     }
 
-    ~DeviceShard() {
+    ~DeviceShard() {  // NOLINT
       dh::safe_cuda(cudaSetDevice(device_));
     }
